Hotel Inspector (part 1)
Added hotel inspector achievement and conversion SQL, icon etc.

Still need to implement game logic and create a click-through which shows the breakdown of all the different methods, when they were obtained with links to the games etc. Will plug this in for checkout completeness too.

Also converted AchievementEntity to Kotlin as I added a new column to it. DB version is now 5.